chore: skeleton for owl-bot responding to new docker images (#1341)
