+++ packages/owl-bot/.eslintignore
